manually scrub url-parse 1.2.0...
This is a downstream dependency from the dated (deprecated) pouch DB
authentication helper plug-in. It contains some critical security
vulnerabilities.

The added resolution bumps this to a more modern version and seems to be
compatible via manual testing. This should be considered a workaround,
pending a better fix of either updating or replacing the pouch DB auth
plugging.